third_party/WebKit/LayoutTests/inspector/device-orientation-success-expected.txt

CONSOLE MESSAGE: line 18: alpha: 20 beta: 30 gamma: 40
CONSOLE MESSAGE: line 18: alpha: 1.1 beta: 2.2 gamma: 3.3
